Pull out some shared functions
