+++ dev/app/Console/Commands/AddBlock.php
@@ -42,18 +42,6 @@ class AddBlock extends Command
     }
 
     /**
-     * Check if a file doesn't already exist.
-     *
-     * @return bool|null
-     */
-    protected function checkExistence($type, $path)
-    {
-        if (File::exists(base_path($path))) {
-            throw new \Exception("{$type} '{$path}' already exists.");
-        }
-    }
-
-    /**
      * Create fieldset.
      *
      * @return bool|null
@@ -81,34 +69,5 @@ class AddBlock extends Command
             ->replace('{{ filename }}', $this->filename);
 
         File::put(base_path("resources/views/page_builder/_{$this->filename}.antlers.html"), $contents);
-    }
-
-    /**
-     * Update page_builder.yaml.
-     *
-     * @return bool|null
-     */
-    protected function updatePageBuilder()
-    {
-        $fieldset = Yaml::parseFile(base_path('resources/fieldsets/page_builder.yaml'));
-        $newSet = [
-            'display' => $this->block_name,
-            'instructions' => $this->instructions,
-            'fields' => [
-                [
-                    'import' => $this->filename
-                ]
-            ]
-        ];
-
-        $existingSets = Arr::get($fieldset, 'fields.0.field.sets');
-        $existingSets[$this->filename] = $newSet;
-        $existingSets = collect($existingSets)->sortBy(function ($value, $key) {
-            return $key;
-        })->all();
-
-        Arr::set($fieldset, 'fields.0.field.sets', $existingSets);
-
-        File::put(base_path('resources/fieldsets/page_builder.yaml'), Yaml::dump($fieldset, 99, 2));
     }
 }
