+++ dev/app/Console/Commands/AddPartial.php
@@ -49,18 +49,6 @@ class AddPartial extends Command
     }
 
     /**
-     * Check if a file doesn't already exist.
-     *
-     * @return bool|null
-     */
-    protected function checkExistence($type, $path)
-    {
-        if (File::exists(base_path($path))) {
-            throw new \Exception("{$type} '{$path}' already exists.");
-        }
-    }
-
-    /**
      * Create partial.
      *
      * @return bool|null
