+++ dev/app/Console/Commands/InstallBlock.php
@@ -77,33 +77,4 @@ class InstallBlock extends Command
         File::put(base_path("resources/fieldsets/{$this->filename}.yaml"), File::get(__DIR__."/stubs/blocks/{$this->filename}.yaml.stub"));
         File::put(base_path("resources/views/page_builder/_{$this->filename}.antlers.html"), File::get(__DIR__."/stubs/blocks/{$this->filename}.antlers.html.stub"));
     }
-
-    /**
-     * Update page_builder.yaml.
-     *
-     * @return bool|null
-     */
-    protected function updatePageBuilder()
-    {
-        $fieldset = Yaml::parseFile(base_path('resources/fieldsets/page_builder.yaml'));
-        $newSet = [
-            'display' => $this->block_name,
-            'instructions' => $this->instructions,
-            'fields' => [
-                [
-                    'import' => $this->filename
-                ]
-            ]
-        ];
-
-        $existingSets = Arr::get($fieldset, 'fields.0.field.sets');
-        $existingSets[$this->filename] = $newSet;
-        $existingSets = collect($existingSets)->sortBy(function ($value, $key) {
-            return $key;
-        })->all();
-
-        Arr::set($fieldset, 'fields.0.field.sets', $existingSets);
-
-        File::put(base_path('resources/fieldsets/page_builder.yaml'), Yaml::dump($fieldset, 99, 2));
-    }
 }
