+++ dev/app/Console/Commands/SharedFunctions.php
@@ -4,8 +4,44 @@ namespace App\Console\Commands;
 
 trait SharedFunctions {
 
-    public function hello()
+    /**
+     * Check if a file doesn't already exist.
+     *
+     * @return bool|null
+     */
+    public function checkExistence($type, $path)
     {
+        if (File::exists(base_path($path))) {
+            throw new \Exception("{$type} '{$path}' already exists.");
+        }
+    }
 
+    /**
+     * Update page_builder.yaml.
+     *
+     * @return bool|null
+     */
+    protected function updatePageBuilder()
+    {
+        $fieldset = Yaml::parseFile(base_path('resources/fieldsets/page_builder.yaml'));
+        $newSet = [
+            'display' => $this->block_name,
+            'instructions' => $this->instructions,
+            'fields' => [
+                [
+                    'import' => $this->filename
+                ]
+            ]
+        ];
+
+        $existingSets = Arr::get($fieldset, 'fields.0.field.sets');
+        $existingSets[$this->filename] = $newSet;
+        $existingSets = collect($existingSets)->sortBy(function ($value, $key) {
+            return $key;
+        })->all();
+
+        Arr::set($fieldset, 'fields.0.field.sets', $existingSets);
+
+        File::put(base_path('resources/fieldsets/page_builder.yaml'), Yaml::dump($fieldset, 99, 2));
     }
 }
